finalize, also change to state
